Fixes events, locations, and tags model/routes

+++ services/hexathons/src/routes/location.ts
@@ -1,6 +1,7 @@
 import { asyncHandler, BadRequestError, checkAbility } from "@api/common";
 import express from "express";
 
+import { EventModel } from "src/models/event";
 import { LocationModel, Location } from "../models/location";
 
 export const locationRoutes = express.Router();
@@ -11,6 +12,14 @@ locationRoutes.route("/").get(
     const locations = await LocationModel.accessibleBy(req.ability);
 
     return res.send(locations);
+  })
+);
+
+locationRoutes.route("/:id").get(
+  checkAbility("read", "Location"),
+  asyncHandler(async (req, res) => {
+    const location = await LocationModel.findById(req.params.id);
+    return res.send(location);
   })
 );
 
@@ -61,6 +70,12 @@ locationRoutes.route("/:id").put(
 locationRoutes.route("/:id").delete(
   checkAbility("delete", "Location"),
   asyncHandler(async (req, res) => {
+    const event = await EventModel.findOne({
+      location: req.params.id,
+    });
+    if (event) {
+      throw new BadRequestError("Event with location still exists. Location cannot be deleted.");
+    }
     await LocationModel.findByIdAndDelete(req.params.id);
     return res.sendStatus(204);
   })
